Fixed Login Screen, minor App.css cleaning
TODO: instead of username for login use email; still need to do major App.css reorganizing/cleaning
