+++ src/CSS_Files/App.css
@@ -37,79 +37,6 @@
   }
 }
 
-/* Login Styles */
-.login-container {
-  min-height: 100vh;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 1rem;
-}
-
-.login-card {
-  background: white;
-  padding: 2rem;
-  border-radius: 12px;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
-  width: 100%;
-  max-width: 400px;
-}
-
-.login-header {
-  text-align: center;
-  margin-bottom: 2rem;
-}
-
-.login-header h2 {
-  color: #333;
-  margin-bottom: 0.5rem;
-}
-
-.login-header p {
-  color: #666;
-  font-size: 14px;
-}
-
-.login-form {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.form-group {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.form-group label {
-  font-weight: 500;
-  color: #333;
-  text-align: left;
-}
-
-.login-btn {
-  background-color: #667eea;
-  color: white;
-  border: none;
-  padding: 12px;
-  border-radius: 6px;
-  font-size: 16px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.login-btn:hover {
-  background-color: #5a6fd8;
-}
-
-.login-demo-note {
-  text-align: center;
-  margin-top: 1rem;
-  font-size: 12px;
-  color: #666;
-}
 
 /* Header Styles */
 .app-header {
@@ -245,26 +172,6 @@ fieldset {
   justify-content: center;
   align-items: center;
   flex-direction: column;
-}
-
-input[type="text"],
-input[type="email"],
-input[type="tel"],
-input[type="date"],
-input[type="password"] {
-  width: 80%;
-  padding: 8px;
-  margin-bottom: 0.8rem;
-  border-radius: 21px;
-  border: 1px solid #ccc;
-}
-
-button {
-  padding: 0.5rem 1rem;
-  margin-right: 1rem;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
 }
 
 .submit-btn {
