+++ src/components/index.js
@@ -1,3 +1,4 @@
+export { default as Login } from './Login'
 export { default as Dashboard } from './Dashboard'
 export { default as ClientDetail } from './ClientDetail'
 export { default as IRACalculator } from './IRACalculator'
